+++ backend/src/HorseRacing.Domain/Entities/Equines/HorseDocument.cs
@@ -5,7 +5,7 @@ namespace HorseRacing.Domain.Entities;
 public class HorseDocument
 {
     public int Id { get; set; }
-    public int HorseId { get; set; }
+    public long HorseId { get; set; }
     public Horse? Horse { get; set; }
     public string DocumentType { get; set; } = string.Empty;
     public string DocumentUrl { get; set; } = string.Empty;
